requirements.txt
# Black & White Video Colorization - Production Dependencies
# Python 3.9+

# Core ML
torch>=2.0.0
torchvision>=0.15.0

# Image/Video processing
opencv-python-headless>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0
Pillow>=10.0.0

# Web UI
gradio>=4.0.0

# Utilities
tqdm>=4.65.0
